Structure OkBuck's KOTLIN_HOME in the structure it expects (#543)
A failed kotlin compile was full of “unable to find kotlin-stdlib.jar
in the Kotlin home directory.” warnings because the directory structure
did not match that of normal kotlin_home directories, even though the
kotlin-stdlib jar existed. To remove these warnings, match the directory
structure it expects.

Also slid in a few extras that will be needed in the future.