Enable jetifier download and seting up tool binary
Did not create jetifier version as an extension, or search for the version on gradle classpath (equals to how kotlin works) because of the required commons-cli and jetifier-standalone jar, which are coupled to this specific version (I've tried with the other approach as well).

If we want the version to be possibly defined for the user, we must find a way to fetch the specific zip with these dependencies based on that version. At least for the version presented on gradle classpath (1.0.0-alpha02), I didn't find any link to download it