[Changed Targets] Invalidate Buck Lint rule if underlying lib changes
One of the inputs to the Bazel rule is the underling src_release target.
However, for Buck, only the raw srcs and resources are passed to the
lint rule. So, if the underlying lib is invalidated due to a transitive
change, Bazel rebuilds the rule but Buck does not.

In this commit, we add logic to invalidate the Buck rule if the Bazel
rule changes. This is to help eliminate extra noise when comparing Buck
and Bazel. This will lead to additional targets being run on Buck, but
Buck is on a deprecation path.