Add custom publishing target to publish to Maven Central
With the sunsetting of OSSRH, the current version of `gradle-maven-publish-plugin`
can no longer successfully publish using `gradlew publish`. The minimum version of
the plugin that can publish to Maven Central requires an upgrade to Gradle 8,
which is a large upgrade from the current Gradle 6 and will require upgrades
to several other plugins as well.

In order to unblock the ability to publish, we implement a custom publishing task
in this commit that uses the same set of credentials to use a POST command to
publish to maven central. This will allow publishing of new upgrades of okbuck
while other upgrades are being scoped.